+++ modules/python_engine/src/include/PyRun.hpp
@@ -17,8 +17,9 @@
 namespace Nelson {
 //=============================================================================
 NLSPYTHON_ENGINE_IMPEXP ArrayOfVector
-PyRun(Interface* io, bool haveEventsLoop, const wstringVector& commands,
-    const wstringVector& outputs, const wstringVector& names, const ArrayOfVector& values);
+PyRun(Interface* io, bool haveEventsLoop, const void* voidPythonObjectHandle,
+    const wstringVector& commands, const wstringVector& outputs, const wstringVector& names,
+    const ArrayOfVector& values);
 //=============================================================================
 }
 //=============================================================================